requirements.txt
diffusers==0.17.1
nats_py==2.3.1
Pillow==9.5.0
torch==2.0.1
transformers==4.30.2
accelerate==0.20.3
cuda-python==12.2
torchvision==0.15.2
